backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
httpx==0.28.1
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
Pillow==11.1.0
PyMuPDF==1.25.3
openpyxl==3.1.5
numpy==2.2.2
